+++ packages/backend/test-rules/userInfoC.test.js
@@ -1,18 +1,17 @@
 /*
 * back-end/test-rules/userInfoC.test.js
 */
-import { strict as assert } from 'assert'
 import { test, expect, describe, beforeAll } from '@jest/globals'
 
 import { collection } from 'firebase-jest-testing/firestoreRules'
 
 // tbd. userinfo likely going to be handled differently (no need for clients to write to it)
 //
-describe.skip("'/userInfo' rules", () => {
+describe("'/userInfo' rules", () => {
   let unauth_userinfoC, abc_userinfoC, def_userinfoC;
 
   beforeAll( () => {
-    const coll = dbAuth.collection('userInfo');   // root collection
+    const coll = collection('userInfo');   // root collection
 
     unauth_userinfoC = coll.as(null);
     abc_userinfoC = coll.as({uid:'abc'});
@@ -22,8 +21,8 @@ describe.skip("'/userInfo' rules", () => {
   //--- UserInfoC read rules ---
 
   test('no-one should be able to read', () => Promise.all([
-      expect( unauth_userinfoC.get("abc") ).toDeny(),   // unauthenticated
-      expect( abc_userinfoC.get("abc") ).toDeny(),     // valid user (cannot read even one's own)
+    expect( unauth_userinfoC.get("abc") ).toDeny(),   // unauthenticated
+    expect( abc_userinfoC.get("abc") ).toDeny(),     // valid user (cannot read even one's own)
   ]));
 
   //--- UserInfoC write (create/update) rules ---